+++ src/FlatNormalGenerator.cpp
@@ -31,7 +31,7 @@ void bmf::FlatNormalGenerator::generate(const std::vector<Triangle>& triangles,
 				continue; // too different
 
 			// almost the same normal => use this vertex
-			outIndices.push_back(i);
+			outIndices.push_back(uint32_t(i));
 			foundVertex = true;
 			break;
 		}
@@ -43,7 +43,7 @@ void bmf::FlatNormalGenerator::generate(const std::vector<Triangle>& triangles,
 		tri.vertex[0].copyAttributesTo(newVertex);
 		newVertex.set(Normal, glm::value_ptr(n));
 
-		outIndices.push_back(outVertices.size());
+		outIndices.push_back(uint32_t(outVertices.size()));
 		outVertices.emplace_back(std::move(newVertex));
 	}
 }
